판매내역 탭 컴포넌트 분리 및 안쓰는 css 정리
